+++ week3/test/test_phylo_codon.py
@@ -30,7 +30,7 @@ def test_upgma():
         [7.0, 7.0, 0.0, 2.0, 4.0],
         [7.0, 6.0, 2.0, 0.0, 3.0],
         [9.0, 8.0, 4.0, 3.0, 0.0],
-    ], dtype=np.float32)
+    ], dtype=np.float64)
     
     tree = upgma(distances)
     newick = tree.to_newick(include_distance=False)
@@ -46,7 +46,7 @@ def test_neighbor_joining():
         [7.0, 7.0, 0.0, 2.0, 4.0],
         [7.0, 6.0, 2.0, 0.0, 3.0],
         [9.0, 8.0, 4.0, 3.0, 0.0],
-    ], dtype=np.float32)
+    ], dtype=np.float64)
     
     tree = neighbor_joining(distances)
     newick = tree.to_newick(include_distance=False)
